requirements.txt
# GeoNode
# from Pypi:
#GeoNode==3.0
-e git+https://github.com/GeoNode/geonode.git#egg=GeoNode
setuptools>=49.2.0
django-markdownify==0.8.1
